+++ packages/nft/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./eventsErc721";
-export * from "./eventsCryptoPunks";
+export * from "./events";
 export * from "./model";
 export * as streams from "./streams";
